+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 ### Added
-- New `popupWithLink` prop.
+- New `popupWithLink` variation.
 ## [2.59.0] - 2021-04-22
 ### Added
 - I18n Jp and No.